requirements.txt
absl-py==1.4.0
amqp==5.1.1
asgiref==3.5.1
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
cachetools==5.2.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.3.0
cryptography==37.0.2
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.0.4
django-allauth==0.50.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.12.0
django-environ==0.9.0
django-filter==21.1
django-gcloud-storage==0.5.0
django-ranged-response==0.2.0
django-redis==5.2.0
django-simple-captcha==0.5.17
django-storages==1.12.3
django-templated-mail==1.1.1
django-timezone-field==5.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.3.0
et-xmlfile==1.1.0
eventlet==0.33.3
exceptiongroup==1.1.1
executing==1.2.0
fake-useragent==1.1.3
flatbuffers==23.5.26
fonttools==4.39.4
gast==0.4.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
greenlet==1.1.2
grpcio==1.54.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.8.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.12.0
ipython==8.14.0
itypes==1.2.0
jax==0.4.12
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.2.4
libclang==16.0.0
loguru==0.7.0
lxml==4.9.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mysql==0.0.3
mysql-connector-python==8.0.29
mysqlclient==2.1.0
numpy==1.22.3
oauthlib==3.2.0
openai==0.25.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.9
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==1.4.2
pandas-stubs==1.5.2.230105
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
postgres==4.0
prompt-toolkit==3.0.38
protobuf==4.23.2
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.15.1
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytube==15.0.0
pytz==2022.1
redis==3.5.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.10.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
sqlparse==0.4.2
stack-data==0.6.2
stripe==3.3.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
termcolor==2.3.0
tqdm==4.64.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
types-pytz==2022.7.1.0
typing_extensions==4.4.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.6
webdriver-manager==3.8.6
Werkzeug==2.3.6
whitenoise==6.2.0
win32-setctime==1.1.0
wincertstore==0.2
wordcloud==1.9.1.1
wrapt==1.14.1
wsproto==1.2.0
zipp==3.8.0
